feat(web): display boundary shapefiles (idas-131)
- refactored to consolidate the modules
