+++ packages/forms-web-app/src/scripts/projects-map/map-setup.js
@@ -1,6 +1,24 @@
 'use strict';
 
+const proj4 = require('proj4/dist/proj4');
+const { register } = require('ol/proj/proj4.js');
+const { get: getProjection } = require('ol/proj.js');
 const { defaults } = require('ol/control/defaults.js');
+const { EPSG_27700, EPSG27700_PROJ4_DEF, EPSG27700_EXTENT } = require('./constants');
+
+/**
+ * Registers the EPSG:27700 (British National Grid) projection with proj4 and OL,
+ * then sets its valid extent and returns the configured OL projection object.
+ *
+ * @returns {import('ol/proj/Projection').default}
+ */
+function setupEpsg27700() {
+	proj4.defs(EPSG_27700, EPSG27700_PROJ4_DEF);
+	register(proj4);
+	const projection = getProjection(EPSG_27700);
+	projection.setExtent(EPSG27700_EXTENT);
+	return projection;
+}
 
 /**
  * Returns OL map controls configured with accessible GOV.UK zoom labels.
@@ -34,7 +52,7 @@ function updateZoomButtons(map) {
 	const view = map.getView();
 	const zoom = view.getZoom();
 	const el = map.getTargetElement();
-	const CSS_CLASS_ZOOM_DISABLED = 'no-zoom'; // Hardcoded here as this is a DOM helper, or can import from constants
+	const CSS_CLASS_ZOOM_DISABLED = 'no-zoom';
 
 	el.querySelector('.ol-zoom-in')?.classList.toggle(
 		CSS_CLASS_ZOOM_DISABLED,
@@ -46,4 +64,4 @@ function updateZoomButtons(map) {
 	);
 }
 
-module.exports = { getControls, updateZoomButtons };
+module.exports = { setupEpsg27700, getControls, updateZoomButtons };
